% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_algorithms.R
\name{combine_algorithms}
\alias{combine_algorithms}
\title{Combine two phylogenetic tree pruning algorithms}
\usage{
combine_algorithms(
  tree_file,
  algorithms,
  total_percent_remaining = 90,
  prune_percentages,
  algorithm_args = list(),
  output = "combined_tree.nwk"
)
}
\arguments{
\item{tree_file}{character; a phylogenetic tree in .nwk, .newick or .tree format.}

\item{algorithms}{character vector of length 2. Allowed values are "CPA", "IQR", and "PSFA".}

\item{total_percent_remaining}{numeric; final percentage of tips to retain after both algorithms.}

\item{prune_percentages}{numeric vector of length 2; percentage of original tips to prune by each algorithm.
For example, c(4, 6) means prune 4 percent with the first algorithm and 6 percent with the second.}

\item{algorithm_args}{named list; optional algorithm-specific arguments.
See all possible arguments in the help section of each argument.}

\item{output}{character; desired name of the final output file.}
}
\value{
A list containing the final tree, percentage retained compared to the original, and step-wise results.
  The step-wise results for the first show the percentage of remaining tips compared to the original.
  For the second step, it shows the percentage of remaining tips compared to the first step.
}
\description{
Runs two tree pruning algorithms sequentially.
}
