% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_tree_CPA.R
\name{prune_tree_CPA}
\alias{prune_tree_CPA}
\title{Prune a Phylogenetic Tree Using the Circular Pruning Algorithm (CPA)}
\usage{
prune_tree_CPA(
  tree_file,
  root_to_node_ratio = 0.1,
  min_num_of_roots = 15,
  M_n = 0,
  threshold = 90,
  beta = 20,
  radius_ratio = 0,
  safe_tips = character(0),
  show_plot = FALSE,
  show_pruned_tips = FALSE,
  output = "cpa_tree.nwk"
)
}
\arguments{
\item{tree_file}{character; a phylogenetic tree in .nwk, .newick or .tree format.}

\item{root_to_node_ratio}{numeric; determines how many nodes does the algorithm try out as "roots".
By increasing this parameter, the precision and the runtime becomes greater too.
The default value means that the algorithm will try out a randomly selected 10% of the non-leaf nodes as "roots".
Apart from these random nodes the algorithm will always try the "midpoint root" of the tree, as it usually gives a good result.
The parameter must be between 0 and 1, although the algorithm can handle the other cases as well (as long as it is a number). Defaults to 0.1.}

\item{min_num_of_roots}{integer; the minimum number of nodes that the algorithm tries out as "roots".
This parameter only matters if the root_to_node_ratio chooses less nodes than min_num_of_roots.
It must be less than the number of non-leaf nodes, although the algorithm can handle the other cases as well (as long as it is an integer).
Defaults to 15.}

\item{M_n}{function; a function, which you should define before calling the algorithm if you intend to use this parameter.
During the CPA the leaves are being placed in brackets based on their distances from the "root".
The algorithm will delete the leaves in the last couple of brackets if certain conditions are met.
M_n determines the number of brackets: the total number of brackets will be M_n(num_leaves), where num_leaves is the number of leaves on the tree.
The default value means that the user did not give any function to the argument. In this case M_n(num_leaves) = num_leaves.
Assigning anything to M_n other than a function will result in an error message.}

\item{threshold}{numeric; the minimum percentage of the leaves that the algorithm must keep. Defaults to 90.}

\item{beta}{numeric; determines where to prune the tree. After placing the leaves in the brackets, the first couple of brackets will be safe, until the total number of leaves in them reaches alpha*num_leaves.
When the CPA finds this threshold, it is ready for pruning. The algorithm will start the pruning if a bracket contains beta percent less leaves than the previous one.
After this point the CPA will delete every leaf. Defaults to 25.}

\item{radius_ratio}{numeric;  controls whether pruning occurs based on the tree's radius. If the longest root-to-tip distance in the pruned tree is not at least radius_ratio percent smaller than in the original tree (measured from the same root), no pruning is performed.
Defaults to 0.}

\item{safe_tips}{list; a list where you can input the names of the tips that you do not want to be cut off.
After the algorithm chose the best root and started the pruning, it checks which leaves are in the list and leaves them on the tree.
The default value means that there are no safe leaves. It is important that the names in the list must be characters.}

\item{show_plot}{bool; the default value means that after running the code it will not show any plot. By setting this value to TRUE, the brackets will be shown to the user.
These brackets correspond to the "best root" (the center of the outscribed circle of the tree). On the x axis you can see the distances from the root, while on the y axis you can see how many leaves were placed in each bracket.
Blue indicates the preserved brackets, while red indicates the ones which have been deleted.}

\item{show_pruned_tips}{bool; responsible for indicating which leaves have been cut from the tree. By setting this value to TRUE, the function will return a list that contains the names of the pruned tips, in addition to the default return values.
This list will be the third return value.}

\item{output}{character; the desired name of the output file. Defaults to "cpa_tree.nwk".}
}
\value{
A list with the following elements:
\describe{
  \item{tree}{The pruned tree in phylo format.}
  \item{percent_remaining}{A numeric value indicating the percentage of leaves retained after pruning.}
  \item{pruned_tips}{(Optional) A character vector containing the names of the pruned tips. This is returned only if `show_pruned_tips = TRUE`.}
}
}
\description{
This function implements a more complex (and often better) algorithm than PSFA but it also has a longer runtime.
It aims to prune the tree to make it roughly circular
}
\details{
You have to define criteria for considering an edge as "excessively long".
You also have to specify the tolerance range.
}
