% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_tree_IQR.R
\name{prune_tree_IQR}
\alias{prune_tree_IQR}
\title{Prune a Phylogenetic Tree Using the IQR Algorithm
This function implements a two-part tree pruning algorithm based on branch length
and root-to-tip distance outlier detection using the Inter Quartile Range (IQR) method.}
\usage{
prune_tree_IQR(tree_file, threshold = 90, output = "iqr_tree.nwk")
}
\arguments{
\item{tree_file}{character; a phylogenetic tree in .nwk, .newick or .tree format.}

\item{threshold}{numeric; the minimum percentage of the leaves that the algorithm must keep. Defaults to 90.}

\item{output}{character; the desired name of the output file. Defaults to "iqr_tree.nwk".}
}
\value{
A list with two elements:
\describe{
  \item{tree}{The pruned tree in phylo format.}
  \item{percent_remaining}{A numeric value indicating the percentage of leaves retained after pruning.}
}
}
\description{
Prune a Phylogenetic Tree Using the IQR Algorithm
This function implements a two-part tree pruning algorithm based on branch length
and root-to-tip distance outlier detection using the Inter Quartile Range (IQR) method.
}
\details{
Abbreviations:

- IQR: the upper fence of the Inter Quartile Range of a vector: Q3 + 3 * (Q3 - Q1) = extreme outlier threshold.
- R2T: root-to-tip distance on a midpoint rooted tree.

First part of the algorithm: Pruning the unrooted tree based on the upper fence of IQR branch lengths.

Calculating the minimum number of tips for each branch (unrooted bifurcating tree, both directions are looked up, than the least tip number is chosen to represent the branch).

Calculating the IQR for branch lengths.

Excluding those extreme outlier branches containing less than 0.05 (or a given) proportion of the tips.

Second part of the algorithm: Pruning the midpoint rooted tree based on root-to-tip distances. Excluding tips based on the R2T IQRs, while iteratively midpoint rooting and excluding the top greatest extreme outlier.

Midpoint rooting the tree (after pruning with method described in the first part).

Calculating the IQR for root-to-tip distances.

Excluding the most extreme outlier tip based on the IQR for root-to-tip distances.

Repeating the previous three steps until there are no more extreme outlier IQR tip is found.

Unrooting the pruned tree.
}
