% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_tree_PSFA.R
\name{prune_tree_PSFA}
\alias{prune_tree_PSFA}
\title{Prune a Phylogenetic Tree Using the Primitive Straight-Forward Approach (PSFA)}
\usage{
prune_tree_PSFA(
  tree_file,
  threshold = 90,
  longest_to_average = 9,
  output = "psfa_tree.nwk"
)
}
\arguments{
\item{tree_file}{character; a phylogenetic tree in .nwk, .newick or .tree format.}

\item{threshold}{numeric; the minimum percentage of the leaves that the algorithm must keep. Defaults to 90.}

\item{longest_to_average}{numeric; determines how many times longer should a branch be than the average, to consider it too long. Defaults to 9.}

\item{output}{character; the desired name of the output file. Defaults to "psfa_tree.nwk".}
}
\value{
A list with two elements:
\describe{
  \item{tree}{The pruned tree in phylo format.}
  \item{percent_remaining}{A numeric value indicating the percentage of leaves retained after pruning.}
}
}
\description{
This functions Identifies the longest edge and calculates the sizes of the two components resulting from its removal.
If the size of the smaller component is within the tolerance range and the edge was excessively long, removes it.
}
\details{
You have to define criteria for considering an edge as "excessively long".
You also have to specify the tolerance range.
}
