% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetune-package.R
\docType{package}
\name{treetune-package}
\alias{treetune}
\alias{treetune-package}
\title{treetune: A package for pruning phylogenetic trees.}
\description{
The treetune package provides algorithms for pruning phylogenetic trees,
including the Circular Pruning Algorithm (CPA),
Interquartile Range algorithm (IQR), Primitive Straight-forward Approach (PSFA)
and sequential combinations of these methods.
}
\details{
The main pruning functions are:
\itemize{
  \item \code{\link{prune_tree_CPA}}: Circular Pruning Algorithm
  \item \code{\link{prune_tree_IQR}}: Interquartile Range algorithm
  \item \code{\link{prune_tree_PSFA}}: Primitive Straight-forward Approach
  \item \code{\link{combine_algorithms}}: Sequentially combines two pruning algorithms.
}
}
\author{
\strong{Maintainer}: Agoston Hunya \email{ago.hunya@gmail.com} (\href{https://orcid.org/<ORCID HERE>}{ORCID})

Authors:
\itemize{
  \item Daniel Hegedus (\href{https://orcid.org/<ORCID HERE>}{ORCID})
  \item Mark Hunor Juhasz (\href{https://orcid.org/<ORCID HERE>}{ORCID})
  \item Tamas Stirling (\href{https://orcid.org/0000-0002-8964-6443}{ORCID})
  \item Balazs Papp (\href{https://orcid.org/0000-0003-3093-8852}{ORCID})
  \item Eszter Ari (\href{https://orcid.org/0000-0001-7774-1067}{ORCID})
}

}
\keyword{internal}
