\name{treetune-package}
\alias{treetune-package}
\alias{treetune}
\docType{package}
\title{
\packageTitle{treetune}
}
\description{
\packageDescription{treetune}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{treetune}

Maintainer: \packageMaintainer{treetune}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = treetune)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{treetune}}
%\section{Documentation Index}{\packageIndices{treetune}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
